// src/app/model/accounts.model.ts
export class Accounts {

  id: number | undefined;
  account_id: string | undefined;
  account_name: string | undefined;
  administrator: string | undefined;
  administrator_id: number | undefined;
  balance: string | undefined;
  capture_id: number | undefined;
  company_id: number | undefined;
  createdAt: string | undefined;
  currency: string | undefined;
  extension: string | undefined;
  month: string | undefined;
  preparer: string | undefined;
  preparer_id: number | undefined;
  reviewer: string | undefined;
  reviewer_id: number | undefined;
  risk: string | undefined;
  status: string | undefined;
  updatedAt: string | undefined;
  workpaper_id: number | undefined;
  year: number | undefined;
  certify: boolean | undefined;
  description: string | undefined;
  captureLogsId: number | undefined;
  action: string | undefined;
  foreignBalance: string | undefined;
}